remove IcebergTable struct, replace with IcebergManifestList
