+++ src/core/metadata/iceberg_table_metadata.cpp
@@ -4,7 +4,6 @@
 
 #include "common/iceberg_utils.hpp"
 #include "catalog/rest/api/catalog_utils.hpp"
-#include "function/metadata/iceberg_metadata.hpp"
 #include "core/metadata/snapshot/iceberg_snapshot.hpp"
 #include "rest_catalog/objects/list.hpp"
 #include "catalog/rest/api/iceberg_create_table_request.hpp"
