+++ src/core/metadata/schema/iceberg_table_schema.cpp
@@ -1,8 +1,6 @@
 #include "core/metadata/schema/iceberg_table_schema.hpp"
 
 #include "duckdb/common/exception.hpp"
-
-#include "function/metadata/iceberg_metadata.hpp"
 #include "common/iceberg_utils.hpp"
 #include "rest_catalog/objects/list.hpp"
