+++ src/include/core/metadata/manifest/iceberg_manifest_list.hpp
@@ -171,6 +171,9 @@ public:
 public:
 	static LogicalType FieldSummaryType();
 	static Value FieldSummaryFieldIds();
+	static unique_ptr<IcebergManifestList> Load(const string &iceberg_path, const IcebergTableMetadata &metadata,
+	                                            const IcebergSnapshot &snapshot, ClientContext &context,
+	                                            const IcebergOptions &options);
 
 private:
 	string path;
